Handle Depost and Withdraw on SavingsModule

+++ src/utils/decimalsToWei.ts
@@ -1,5 +1,5 @@
 import { BigInt } from "@graphprotocol/graph-ts";
 
-export function decimalsToWei(decimals: u8): BigInt {
-  return BigInt.fromI32(10).pow(decimals);
+export function decimalsToWei(decimals: f64): BigInt {
+  return BigInt.fromI32(10).pow(decimals as u8);
 }
